Broke up app.js into routes, passport, and middleware
Broke up app.js into modules.
Updated folder structure and renamed config to auth. Removed
unnecessary variables from app.js.
