+++ config/auth.js
@@ -1,0 +1,13 @@
+var auth = {};
+
+auth.twitter = {};
+
+
+auth.twitter.consumerKey 			= "uf6Ib1NWNAt0KiILE9f2O0qNk";
+auth.twitter.consumerSecret 		= "4ZL44FOaCqbI2QwR9z8Y76EN6Y87oyhsDABU7fC1MtO3habUCi";
+auth.twitter.accessTokenKey 		= "1564043540-wnKSh6DJNaJmuMBKqhSOXMJH3BEndrgDtSdjbXF";
+auth.twitter.accessTokenSecret 	= "NsofwoBnfZ4r1gpnbj2bhR09NNPa196MjGT0wSwAo0RuN";
+auth.twitter.callbackURL			= "http://127.0.0.1:3000/login/twitter/callback"
+
+// export config for inclusion in app.js
+module.exports = auth;
